modified:   create-live-menus.sh 	modified:   get-torrents.sh

+++ create-live-menus.sh
@@ -5,7 +5,7 @@ scriptdir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 source $scriptdir/pxe.config
 
 nfs_path=$nfs_server:/pxeboot
-nfs_root_path=$nfs_server:/var/nfs/pxeboot/live
+nfs_root_path=$nfs_server:/var/nfs/pxeboot/stock
 seed_file="ubuntu.seed"
 
 function check_for_sudo ()
@@ -31,29 +31,29 @@ EOM
 function live_casper_initrd_lz ()
 {
 kernelpath=$bootfolder/casper
-mkdir -p $tftp_folder/$kernelpath
-cp -uv $subfolder/casper/vmlinuz $tftp_folder/$kernelpath/
-cp -uv $subfolder/casper/initrd.lz $tftp_folder/$kernelpath/
-cat >> $menupath << EOM
-LABEL $revdate
-MENU LABEL $revdate
-    kernel $kernelpath/vmlinuz
-    append initrd=$kernelpath/initrd.lz noprompt boot=casper url=$seed_path/$seed_file netboot=nfs nfsroot=$nfs_root_path/$distro/$revision ro toram -
+if [ -f $subfolder/casper/vmlinuz ]; then
+	mkdir -p $tftp_folder/$kernelpath
+	cp -uv $subfolder/casper/vmlinuz $tftp_folder/$kernelpath/
+	cp -uv $subfolder/casper/initrd.lz $tftp_folder/$kernelpath/
+	cat >> $menupath << EOM
+	LABEL $revdate
+	MENU LABEL $revdate
+	    kernel $kernelpath/vmlinuz
+	    append initrd=$kernelpath/initrd.lz noprompt boot=casper url=$seed_path/$seed_file netboot=nfs nfsroot=$nfs_root_path/$distro/$revision ro toram -
 EOM
-}
-
-function live_casper_initrd_gz ()
-{
-kernelpath=$bootfolder/casper
-mkdir -p $tftp_folder/$kernelpath
-cp -uv $subfolder/casper/vmlinuz $tftp_folder/$kernelpath/
-cp -uv $subfolder/casper/initrd.gz $tftp_folder/$kernelpath/
-cat >> $menupath << EOM
-LABEL $revdate
-MENU LABEL $revdate
-    kernel $kernelpath/vmlinuz
-    append initrd=$kernelpath/initrd.gz noprompt boot=casper url=$seed_path/$seed_file netboot=nfs nfsroot=$nfs_root_path/$distro/$revision ro toram -
+elif [ -f $subfolder/casper/vmlinuz.efi ]; then
+	mkdir -p $tftp_folder/$kernelpath
+	cp -uv $subfolder/casper/vmlinuz.efi $tftp_folder/$kernelpath/
+	cp -uv $subfolder/casper/initrd.lz $tftp_folder/$kernelpath/
+	cat >> $menupath << EOM
+	LABEL $revdate
+	MENU LABEL $revdate
+	    kernel $kernelpath/vmlinuz.efi
+	    append initrd=$kernelpath/initrd.lz noprompt boot=casper url=$seed_path/$seed_file netboot=nfs nfsroot=$nfs_root_path/$distro/$revision ro toram -
 EOM
+else
+	break
+fi
 }
 
 function generate_live_menu ()
@@ -78,8 +78,6 @@ for folder in /mnt/stock/*; do
             bootfolder=boot/$distro/$revision
             if [ -e "$subfolder/casper/initrd.lz" ]; then
                 live_casper_initrd_lz
-            elif [ -e "$subfolder/casper/initrd.gz" ]; then
-                live_casper_initrd_gz
             else
                 echo " not a live cd"
             rm $menupath
